requirements.txt
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0

# asgiref==3.7.2
# black==24.2.0
# certifi==2024.2.2
# cffi==1.16.0
# charset-normalizer==3.3.2
# click==8.1.7
# colorama==0.4.6
# cryptography==42.0.5
# defusedxml==0.8.0rc2
# django-templated-mail==1.1.1
# djangorestframework-simplejwt==5.3.1
# djoser==2.2.2
# flake8==7.0.0
# idna==3.6
# mccabe==0.7.0
# mypy-extensions==1.0.0
# oauthlib==3.2.2
# packaging==23.2
# pathspec==0.12.1
# platformdirs==4.2.0
# pycodestyle==2.11.1
# pycparser==2.21
# pyflakes==3.2.0
# PyJWT==2.8.0
# python3-openid==3.2.0
# pytz==2024.1
# requests==2.31.0
# requests-oauthlib==1.3.1
# social-auth-app-django==5.4.0
# social-auth-core==4.5.3
# sqlparse==0.4.4
# tzdata==2024.1
# urllib3==2.2.1
